+++ spec/dummy/app/controllers/admin/releases_controller.rb
@@ -8,4 +8,12 @@ class Admin::ReleasesController < Fae::BaseController
     @item.build_label_pdf if @item.label_pdf.blank?
   end
 
+  def attributes_for_cloning
+    [:name, :slug, :intro, :body, :wine_id, :release_date]
+  end
+
+  def associations_for_cloning
+    [:aromas, :events]
+  end
+
 end